telegram/requirements.txt
aiogram==2.9.2
fastapi==0.100.0
pydantic_settings==2.4.0
uvicorn==0.29.0
loguru==0.7.0
python-dotenv~=1.0.1
